% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GenerateAccessToken}
\alias{GenerateAccessToken}
\title{Gets an OAuth 2.0 Access Token by authorizing the user account to the Google
Analytics API}
\usage{
GenerateAccessToken(secrets.file.path = NULL)
}
\description{
Gets an OAuth 2.0 Access Token by authorizing the user account to the Google
Analytics API
}
\details{
When evaluated for the first time this function asks for User Consent
for the Google Analytics Account and retrieves the Access and Refresh Tokens
for Authorization. These tokens are saved locally to a file on the user's system.
If the user had authorized an account earlier and refresh token is already found
on the user's system, then this function retrives a new Access Token and updates
the Access Token File in user's memory.
}

