\name{GetAcctDataFeedJSON}

\alias{GetAcctDataFeedJSON}

\title{A function to capture the profile JSON data into an R data...}

\usage{	
This query.uri is prepared query string for GA account data retrival will be used as input argument and it will return the api.response.json with profile ids and profile names. 
	 
   api.response.json <- GetAcctDataFeedJSON(query.uri)
   
   }

\description{This function will request to Google Management API with the query prepared by the QueryBuilder class for retriving the GA Account data.}

\arguments{query.uri: The data feed query strig generated by the Query builder Class}

\value{GA.Data: the response as account data feed in the JSON format.}