% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetAppCredentials}
\alias{GetAppCredentials}
\title{Save App Credentials (Client ID and Client Secret) locally to system}
\usage{
GetAppCredentials(client.id, client.secret)
}
\arguments{
\item{client.id}{Client ID of the Application}

\item{client.secret}{Client Secret of the Application}
}
\value{
Saves the App Credentials to a file on the user's system
}
\description{
This function gets the Client ID and Client Secret of the Application from
the user and saves it locally to a file for OAuth 2.0 Authorization flow.
}
\author{
Kushan Shah
}

