% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GetDataFeed.R
\name{GetDataFeed}
\alias{GetDataFeed}
\title{This will request with the prepared Query to the Google Analytics
Data feed API and returns the data in dataframe R object.}
\usage{
GetDataFeed(query.uri, caching.dir = NULL, caching = FALSE)
}
\arguments{
\item{query.uri}{The URI prepared by the QueryBuilder class.}

\item{caching.dir}{String Directory to save cached data}

\item{caching}{Boolean caching required?}
}
\value{
GA.list The Google Analytics API JSON response converted to a
list object
}
\description{
This will request with the prepared Query to the Google Analytics
Data feed API and returns the data in dataframe R object.
}
\keyword{internal}

