% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetProfileData}
\alias{GetProfileData}
\title{Retrieves the list of Profiles for the Google Analytics Account}
\usage{
GetProfileData()
}
\value{
profiles R dataframe with profile id and profile name.
}
\description{
This function will retrive the available profiles from your
Google analytics account by the Google Management API with the help of
the access token.
}

