% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetProfiles}
\alias{GetProfiles}
\title{Retrieves the list of Profiles for the Google Analytics Account

This function retrieves all the available profiles from the
Google analytics account which has been authenticated by the user
Before retrieving the list of profiles, this function checks the token
for expiry}
\usage{
GetProfiles(token)
}
\arguments{
\item{token}{Token Object generated by \code{Auth()}}
}
\value{
profiles R dataframe with profile id and profile name.
}
\description{
Retrieves the list of Profiles for the Google Analytics Account

This function retrieves all the available profiles from the
Google analytics account which has been authenticated by the user
Before retrieving the list of profiles, this function checks the token
for expiry
}

