% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GetProfilesFromJSON.R
\name{GetProfilesFromJSON}
\alias{GetProfilesFromJSON}
\title{This function will do the parsing operation on the JSON reponse
returned from the Google Management API and return the
dataframe stored with the profile id and profile name}
\usage{
GetProfilesFromJSON(api.response.json)
}
\arguments{
\item{api.reponse.json}{The JSON response from GetProfileData function which will request to the Google Management API.}
}
\value{
Profileres.list The list stored with totalResults as value of the total available data rows and profiles as the R dataframe object with two columns as column id and column name.
}
\description{
This function will do the parsing operation on the JSON reponse
returned from the Google Management API and return the
dataframe stored with the profile id and profile name
}
\keyword{internal}

