% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetReportData}
\alias{GetReportData}
\title{Queries the Google Analytics API for the specified dimensions,metrics and other query parameters}
\usage{
GetReportData(query.builder, split_daywise = FALSE, paginate_query = FALSE)
}
\arguments{
\item{query.builder}{Name of the object corresponding to the QueryBuilder class}

\item{paginate_query}{Pages through chunks of results by requesting maximum
number of allowed rows at a time. Default Value of this argument is FALSE. Note that
if this parameter is set to True, queries will take more longer to complete and use
more Quota. For more on Google Analytics API quota check
https://developers.google.com/analytics/devguides/reporting/core/v3/limits-quotas#core_reporting}

\item{split_daywise}{Splits the query by date range into sub queries of
single days. Default value of this argument is FALSE. Setting this
argument to true automatically paginates through each daywise query. Note that
if this parameter is set to True, queries will take more longer to complete and use
more Quota}
}
\value{
api.response dataframe containing the response of the Google Analytics API
}
\description{
This function will retrieve the data by firing the query to the Core Reporting API. It also displays
status messages after the completion of the query. The user also has the option split the query into
daywise partitions and paginate the query responses in order to decrease the effect the sampling
}
\examples{
\dontrun{
ga.df <- GetReportData(query)
ga.df <- GetReportData(query,split_daywise=True)
ga.df <- GetReportData(query,paginate_query=True)
}
}
\seealso{
Queries can be tested in the Google Analytics Query Feed Explorer \url{http://ga-dev-tools.appspot.com/explorer/}
}

