% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Init}
\alias{Init}
\title{A function setting initial values of a GA URI query.}
\usage{
Init(start.date = NULL, end.date = NULL, dimensions = NULL,
  metrics = NULL, segment = NULL, sort = NULL, filters = NULL,
  max.results = NULL, start.index = NULL, table.id = NULL)
}
\arguments{
\item{start.date}{See QueryBuilder()}

\item{end.date}{See QueryBuilder()}

\item{dimensions}{See QueryBuilder()}

\item{metrics}{See QueryBuilder()}

\item{segment}{See QueryBuilder()}

\item{sort}{See QueryBuilder()}

\item{filters}{See QueryBuilder()}

\item{max.results}{See QueryBuilder()}

\item{start.index:}{See QueryBuilder()}

\item{table.id:}{See QueryBuilder()}

\item{access_token:}{See AccessToken()}
}
\value{
None Sets the initial query parameters.
}
\description{
A function setting initial values of a GA URI query.
}

