% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ParseApiErrorMessage.R
\name{ParseApiErrorMessage}
\alias{ParseApiErrorMessage}
\title{To check whether the returned JSON response is error or not.
If it is error then it will}
\usage{
ParseApiErrorMessage(api.response.json)
}
\arguments{
\item{api.response.json}{The json data as reposnse returned by the Google Data feed API or Google Management API}
}
\description{
If there is an error in JSON response then this function will return the related error code and message for that error.
}
\keyword{internal}

