% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ParseDataFeedJSON.R
\name{ParseDataFeedJSON}
\alias{ParseDataFeedJSON}
\title{This function will parse the json response and checks if the reponse
is contains an error, if found it will promt user with the related
error message.}
\usage{
ParseDataFeedJSON(GA.Data)
}
\arguments{
\item{GA.Data}{The json reponse returned by the Google analytics Data feed API.}
}
\value{
GA.list.param GA.list.param list object obtained from this json argument GA.Data.
}
\description{
This function will parse the json response and checks if the reponse
is contains an error, if found it will promt user with the related
error message.
}
\keyword{internal}

