% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RefreshToAccessToken}
\alias{RefreshToAccessToken}
\title{This function takes the Refresh Token as an argument and retrives a new
Access Token based on it}
\usage{
RefreshToAccessToken(refresh.token, client.id, client.secret)
}
\arguments{
\item{refresh.token}{Refresh Token that was saved to the local file}

\item{client.id}{Client ID of the Application. This is a OAuth2.0 Credential}

\item{client.secret}{Client Secret of the Application. Again this too is an
                  OAuth2.0 Credential}
}
\value{
access token   New Access Token
}
\description{
This function takes the Refresh Token as an argument and retrives a new
Access Token based on it
}
\keyword{:}
\keyword{Reference}
\keyword{https://developers.google.com/accounts/docs/OAuth2#installed}
\keyword{internal}

