% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RemoveToken}
\alias{RemoveToken}
\title{Deletes the stored Access and Refresh Tokens from the local file}
\usage{
RemoveToken()
}
\description{
In case if the user wants to query a new profile, then the Authorization flow
has to repeated. This requires deleting the stored Access and Refresh Tokens
from the system file
}
\examples{
\dontrun{RemoveToken()}
}
\author{
Vignesh Prajapati
}

