\name{RGoogleAnalytics}
\alias{RGoogleAnalytics}
\alias{RGoogleAnalytics}
\docType{package}
\title{
Provides functions for accessing and retrieving data from the Google Analytics API
}
\description{
RGoogleAnalytics is a wrapper around the Google Analytics Core Reporting API. It provides easy to use functions for authenticating and extracting data from Google Analytics.
}
\details{
\tabular{ll}{
Package: \tab RGoogleAnalytics\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-05-14\cr
License: \tab Apache License 2.0 \cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Michael Pearmain, Nick Mihailowski, Vignesh Prajapati

Maintainer: Kushan Shah <kushan@tatvic.com>
}
\references{
Google Analytics Core Reporting API Reference Guide
https://developers.google.com/analytics/devguides/reporting/core/v3/reference
List of Valid Dimension/Metric Combinations
https://developers.google.com/analytics/devguides/reporting/core/dimsmets
A Tool to Test your queries
http://ga-dev-tools.appspot.com/explorer/
}

\keyword{ package }
\seealso{

}
\examples{
}
