% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RGoogleAnalytics}
\alias{RGoogleAnalytics}
\title{Creates a skeleton shell for accessing the Google Analytics API.}
\usage{
RGoogleAnalytics()
}
\value{
Returns a list of methods, for accessing the Google Analytics Core Reporting API
      GetProfileData()
      GetReportData()
      GetAppCredentials()
      RemoveToken()
      RemoveAppCredentials()
      ValidateToken()
      GenerateAccessToken()
      RefreshToAccessToken()
}
\description{
Creates a skeleton shell for accessing the Google Analytics API.
}
\author{
Michael Pearmain
}

