% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SetColDataType.R
\name{SetColDataType}
\alias{SetColDataType}
\title{This will set the appropriate data type to the each column of the
provided dataframe.}
\usage{
SetColDataType(col.datatype, col.name, dataframe.param)
}
\arguments{
\item{col.datatype}{The vector of the datatype of all the dimensions and metrics from the parsed list data.}

\item{col.name}{The vector of the name of the all dimensions and metrics of the retrived data and it will be attached to the dataframe.param.}

\item{dataframe.param}{}
}
\value{
dataframe.param The dataframe will be set with its column names with the appropriate class type.
}
\description{
This will set the appropriate data type to the each column of the
provided dataframe.
}
\keyword{internal}

