% Generated by roxygen2 (4.0.2): do not edit by hand
\name{SetDataFrameMCF}
\alias{SetDataFrameMCF}
\title{To prepare the dataframe by applying the column names and column datatypes
to the provided data frame.}
\usage{
SetDataFrameMCF(GA.list.param.columnHeaders, dataframe.param)
}
\arguments{
\item{GA.list.param.columnHeaders}{list includes GA response data, column name, column datatype.}

\item{dataframe.param}{The reponse data(dimensions and metrics data rows) packed in dataframe without the column names and column types.}
}
\value{
dataframe.param The dataframe attached with the column  names and column datatype as per type the Dimensions and metrics
}
\description{
To prepare the dataframe by applying the column names and column datatypes
to the provided data frame.
}
\keyword{internal}

