% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SplitQueryDaywise.R
\name{SplitQueryDaywise}
\alias{SplitQueryDaywise}
\title{This function breaks up the time range (as specified by Start Date and End Date) into single days
and hits a query for each day. The responses for each queries are collated into a single dataframe and
returned. This procedure helps decrease the effect of sampling.}
\usage{
SplitQueryDaywise(query.builder, token, delay)
}
\arguments{
\item{query.builder}{Name of the object corresponding to the QueryBuilder class}

\item{token}{Token Object created by the Auth() function}

\item{delay}{Time delay in seconds between successive queries in order to avoid Rate Limit Error}
}
\value{
list containing the Column Headers and the Collated dataframe that represents the query response
}
\description{
This function breaks up the time range (as specified by Start Date and End Date) into single days
and hits a query for each day. The responses for each queries are collated into a single dataframe and
returned. This procedure helps decrease the effect of sampling.
}
\seealso{
https://developers.google.com/analytics/devguides/reporting/core/v2/gdataReferenceDataFeed#largeDataResults
}
\keyword{internal}

