% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ToUri.R
\name{ToUri}
\alias{ToUri}
\title{Returns the URI constructed from the parameter settings. This also
URI-encodes all the values in each query parameter.}
\usage{
ToUri(query.builder, token)
}
\arguments{
\item{query.builder}{Name of the Object of the Query Builder Class}

\item{token}{Token object containing the OAuth2.0 Authentication details}
}
\value{
A full URI that can be used with the Google Analytics API.
}
\description{
Returns the URI constructed from the parameter settings. This also
URI-encodes all the values in each query parameter.
}
\keyword{internal}

