% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ToUriMCF}
\alias{ToUriMCF}
\title{Returns the URI constructed from the parameter settings for the Multi Channel Funnel Reporting API. This also
URI-encodes all the values in each query parameter.}
\usage{
ToUriMCF(query.builder, token)
}
\arguments{
\item{query.builder}{Name of the Object of the Query Builder Class}

\item{token}{Token object containing the OAuth2.0 Authentication details}
}
\value{
A full URI that can be used with the Google Analytics API.
}
\description{
Returns the URI constructed from the parameter settings for the Multi Channel Funnel Reporting API. This also
URI-encodes all the values in each query parameter.
}
\keyword{internal}

