% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ValidateToken}
\alias{ValidateToken}
\title{This function checks whether the Access Token stored in the local file is
expired. If yes, it generates a new Access Token and updates the local file.}
\usage{
ValidateToken()
}
\description{
This function checks whether the Access Token stored in the local file is
expired. If yes, it generates a new Access Token and updates the local file.
}

