% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Auth}
\alias{Auth}
\title{Authorizes the RGoogleAnalytics library to the user's Google Analytics Account}
\usage{
Auth(secrets.file.path = NULL)
}
\arguments{
\item{secrets.file.path}{Optional. Expects a full file path to the client_secrets.json file.
For eg. /home/kushan/Desktop/client_secrets.json. In case if this argument is left blank, then the
function searches for the client_secrets.json file in the RGoogleAnalytics installation directory}
}
\description{
This function expects a client_secrets.json file. In order to obtain this file
}
\details{
- Go to https://console.developers.google.com
- Create a New Project and enable the Google Analytics API
- Select the Application Type "OAuth 2.0 for Installed Applications"
- Download the JSON file and rename it to client_secrets.json
- Copy it to RGoogleAnalytics directory in your R installation or any other custom directory
- The RGoogleAnalytics installation directory can be found by running the command file.path(path.package("RGoogleAnalytics"))
- In case of a custom directory specify the full path in the secrets.file.path argument

When evaluated for the first time this function asks for User Consent
for the Google Analytics Account and retrieves the Access and Refresh Tokens
for Authorization. These tokens are saved locally to a file on the user's system.
In subsequent executions, a browser redirect is not required. This function will automatically
get a New Access Token each time when called. Note that an Access Token is valid for 60 minutes
}

