% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetDataFeed}
\alias{GetDataFeed}
\title{This will request with the prepared Query to the Google Analytics
Data feed API and returns the data in dataframe R object.}
\usage{
GetDataFeed(query.uri)
}
\arguments{
\item{query.uri}{The URI prepared by the QueryBuilder class.}
}
\value{
GA.list The Google Analytics API JSON response converted to a
list object
}
\description{
This will request with the prepared Query to the Google Analytics
Data feed API and returns the data in dataframe R object.
}
\keyword{internal}

