% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetProfiles}
\alias{GetProfiles}
\title{Retrieves the list of Profiles for the Google Analytics Account}
\usage{
GetProfiles()
}
\value{
profiles R dataframe with profile id and profile name.
}
\description{
This function retrives all the available profiles from the
Google analytics account which has been authenticated by the user
}
\examples{
\dontrun{GetProfiles()}
}

