% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetProfiles}
\alias{GetProfiles}
\title{Retrieve the list of Profiles for the Google Analytics Account}
\usage{
GetProfiles(token)
}
\arguments{
\item{token}{Token Object generated by \code{\link{Auth}}}
}
\value{
R dataframe with profile ID and profile name.
}
\description{
This function retrieves all the available profiles from the
Google analytics account.
Before retrieving the list of profiles, this function checks whether the token
is valid and regenerates it if required
}

