% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetReportData}
\alias{GetReportData}
\title{Queries the Google Analytics API for the specified dimensions,metrics and other query parameters}
\usage{
GetReportData(query.builder, split_daywise = FALSE, paginate_query = FALSE)
}
\arguments{
\item{query.builder}{Name of the object corresponding to the QueryBuilder class}

\item{paginate_query}{Pages through chunks of results by requesting maximum
number of allowed rows at a time. Default Value of this argument is FALSE}

\item{split_daywise}{Splits the query by date range into sub queries of
single days. Default value of this argument is FALSE. Note that setting this
argument to true automatically paginates through each daywise query}
}
\value{
api.response The respose is in the dataframe format as the output data returned from the Google Analytics Data feed API.
}
\description{
This function will retrieve the data by firing the query to the Core Reporting API. It also displays
status messages after the completion of the query. The user also has the option split the query into
daywise partitions and paginate the query responses in order to decrease the effect the sampling
}
\examples{
\dontrun{ga_df <- GetReportData(query)
ga_df <- GetReportData(query,split_daywise=True)
ga_df <- GetReportData(query,paginate_query=True)
}
}

