% Generated by roxygen2 (4.0.1): do not edit by hand
\name{QueryBuilder}
\alias{QueryBuilder}
\title{Builds a query with the specified dimensions,metrics and other query parameter}
\usage{
QueryBuilder(query.params.list)
}
\arguments{
\item{query.params.list}{List of all the Query Parameters. See \code{Init} for the
entire list}
}
\value{
builder The builder method to process the query parameters.
}
\description{
Builds a query with the specified dimensions,metrics and other query parameter
}

