% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/QueryBuilder.R
\name{QueryBuilder}
\alias{QueryBuilder}
\title{Initialize a QueryBuilder object with the given parameters and perform validation}
\usage{
QueryBuilder(query.params.list)
}
\arguments{
\item{query.params.list}{List of all the Query Parameters. See \code{\link{Init}} for the
entire list}
}
\value{
The builder object to process the query parameters.
}
\description{
Initialize a QueryBuilder object with the given parameters and perform validation
}

