% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RemoveAppCredentials}
\alias{RemoveAppCredentials}
\title{Deletes App Credentials(Client ID and Client Secret) from the system}
\usage{
RemoveAppCredentials()
}
\arguments{
\item{None}{}
}
\value{
None Deletes the app_credentials.rda file from the user's system
}
\description{
In case if the user creates a new project in the Google API Developer Console, then
a fresh set of OAuth2.0 credentials (Client ID and Client Secret) are provided
This requires deletion of old stored credentials
}
\examples{
\dontrun{RemoveAppCredentials()}
}
\author{
Kushan Shah
}

