% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RemoveToken}
\alias{RemoveToken}
\title{Deletes the stored Access and Refresh Tokens from the system}
\usage{
RemoveToken()
}
\description{
The OAuth 2.0 credentials are account specific. If the user wants to query
a different Google Analytics Account than the one previously authenticated
the older credentials need to be deleted from the system. This function first
searches for the credentials file in the system and deletes the file if found
}

